adding-table and some more
Adding the table to make it easy for user to add the pet from the feature file if required.
Made the @delete_all_pets tag work even if the pet has been deleted in the scenario.
Make the status enum
And seperated the endpoints to live in another class as an enum for it to be easily accessable and reusable.
Added an action class to update the pet id,name,status as provided from the scenario.
Added an overload method to getpet to use it in the delete_all pets.
Cleaned the context and Status class