oneLessConfig: Minimizing config setup
Now we don't have to set
Browser_Under_Test=chrome;Browser_Endpoint=http://localhost:4444/wd/hub
in Environment variable of config file.
It will default to the default value if nothing is specifies.
As the default values are something where the node is trigerred or the browser is user.

Also added Utilization of DevTools